feat: Move "controller" sub-command to "controller run"
