+++ cmd/kluctl/commands/root.go
@@ -66,7 +66,7 @@ type cli struct {
 	Render      renderCmd      `cmd:"" help:"Renders all resources and configuration files"`
 	Seal        sealCmd        `cmd:"" help:"Seal secrets based on target's sealingConfig"`
 	Validate    validateCmd    `cmd:"" help:"Validates the already deployed deployment"`
-	Controller  controllerCmd  `cmd:"" help:"Run the Kluctl controller"`
+	Controller  controllerCmd  `cmd:"" help:"Kluctl controller sub-commands"`
 
 	Version versionCmd `cmd:"" help:"Print kluctl version"`
 }
